# api/openapi.yaml
openapi: 3.0.1
info:
  description: "TODO: Add a description"
  title: FIWARE-NGSI v2 Specification
  version: "1.0"
servers:
- url: http://orion.lab.fiware.org/
paths:
...
  /v2/entities/{entityId}/attrs/{attrName}/value:
    get:
      description: |-
        This operation returns the `value` property with the value of the attribute.
        * If attribute value is JSON Array or Object:
          * If `Accept` header can be expanded to `application/json` or `text/plain` return the value as a JSON with a
            response type of application/json or text/plain (whichever is the first in `Accept` header or
            `application/json` in case of `Accept: */*`).
          * Else return a HTTP error "406 Not Acceptable: accepted MIME types: application/json, text/plain"
        * If attribute value is a string, number, null or boolean:
          * If `Accept` header can be expanded to text/plain return the value as text. In case of a string, citation
            marks are used at the begining and end.
          * Else return a HTTP error "406 Not Acceptable: accepted MIME types: text/plain"
        Response:
        * Successful operation uses 200 OK.
        * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
          more details.
      operationId: Get Attribute Value
      parameters:
      - description: Id of the entity in question
        in: path
        name: entityId
        required: true
        schema:
          type: string
      - description: Name of the attribute to be retrieved.
        in: path
        name: attrName
        required: true
        schema:
          type: string
      - description: |-
          Entity type, to avoid ambiguity in case there are several
          entities with the same entity id.
        in: query
        name: type
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              example:
                address: Ronda de la Comunicacion s/n
                zipCode: 28050
                city: Madrid
                country: Spain
              schema:
                $ref: '#/components/schemas/GetAttributeValueResponse'
          description: ""
      summary: Get Attribute Value
      tags:
      - Attribute Value
      x-unitTests:
      - request:
          method: GET
          uri: /v2/entities/Bcn_Welt/attrs/address/value
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: RAW
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers:
            Content-Type: application/json
          body: "{  \"address\": \"Ronda de la Comunicacion s/n\",  \"zipCode\": 28050,\
            \  \"city\": \"Madrid\",  \"country\": \"Spain\"}"
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Get Attribute Value1
        x-testDescription: |-
          This operation returns the `value` property with the value of the attribute.
          * If attribute value is JSON Array or Object:
            * If `Accept` header can be expanded to `application/json` or `text/plain` return the value as a JSON with a
              response type of application/json or text/plain (whichever is the first in `Accept` header or
              `application/json` in case of `Accept: */*`).
            * Else return a HTTP error "406 Not Acceptable: accepted MIME types: application/json, text/plain"
          * If attribute value is a string, number, null or boolean:
            * If `Accept` header can be expanded to text/plain return the value as text. In case of a string, citation
              marks are used at the begining and end.
            * Else return a HTTP error "406 Not Acceptable: accepted MIME types: text/plain"
          Response:
          * Successful operation uses 200 OK.
          * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
            more details.
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-accepts:
      - application/json
    put:
      description: |-
        The request payload is the new attribute value.
        * If the request payload MIME type is `application/json`, then the value of the attribute is set to
          the JSON object or array coded in the payload (if the payload is not a valid JSON document,
          then an error is returned).
        * If the request payload MIME type is `text/plain`, then the following algorithm is applied to the
          payload:
          * If the payload starts and ends with citation-marks (`"`), the value is taken as a string
            (the citation marks themselves are not considered part of the string)
          * If `true` or `false`, the value is taken as a boolean.
          * If `null`, the value is taken as null.
          * If these first three tests 'fail', the text is interpreted as a number.
          * If not a valid number, then an error is returned and the attribute's value is unchanged.
        The payload MIME type in the request is specified in the `Content-Type` HTTP header.
        Response:
        * Successful operation uses 204 No Content
        * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
          more details.
      operationId: Update Attribute Value
      parameters:
      - description: Id of the entity to be updated.
        in: path
        name: entityId
        required: true
        schema:
          type: string
      - description: Attribute name.
        in: path
        name: attrName
        required: true
        schema:
          type: string
      - in: header
        name: Content-Type
        required: true
        schema:
          type: string
      - description: |-
          Entity type, to avoid ambiguity in case there are several
          entities with the same entity id.
        in: query
        name: type
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAttributeValueRequest'
        required: true
      responses:
        "200":
          content: {}
          description: ""
      summary: Update Attribute Value
      tags:
      - Attribute Value
      x-unitTests:
      - request:
          method: PUT
          uri: /v2/entities/Bcn_Welt/attrs/address/value
          headers:
            Content-Type: application/json
          body: "{  \"address\": \"Ronda de la Comunicacion s/n\",  \"zipCode\": 28050,\
            \  \"city\": \"Madrid\",  \"country\": \"Spain\"}"
        expectedResponse:
          x-allowExtraHeaders: true
          x-bodyMatchMode: NONE
          x-arrayOrderedMatching: false
          x-arrayCheckCount: false
          x-matchResponseSchema: true
          headers: {}
        x-testShouldPass: true
        x-testEnabled: true
        x-testName: Update Attribute Value1
        x-testDescription: |-
          The request payload is the new attribute value.
          * If the request payload MIME type is `application/json`, then the value of the attribute is set to
            the JSON object or array coded in the payload (if the payload is not a valid JSON document,
            then an error is returned).
          * If the request payload MIME type is `text/plain`, then the following algorithm is applied to the
            payload:
            * If the payload starts and ends with citation-marks (`"`), the value is taken as a string
              (the citation marks themselves are not considered part of the string)
            * If `true` or `false`, the value is taken as a boolean.
            * If `null`, the value is taken as null.
            * If these first three tests 'fail', the text is interpreted as a number.
            * If not a valid number, then an error is returned and the attribute's value is unchanged.
          The payload MIME type in the request is specified in the `Content-Type` HTTP header.
          Response:
          * Successful operation uses 204 No Content
          * Errors use a non-2xx and (optionally) an error payload. See subsection on "Error Responses" for
            more details.
      x-operation-settings:
        CollectParameters: false
        AllowDynamicQueryParameters: false
        AllowDynamicFormParameters: false
        IsMultiContentStreaming: false
      x-codegen-request-body-name: body
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
  schemas:
...
    GetAttributeValueResponse:
      example:
        address: Ronda de la Comunicacion s/n
        zipCode: 28050
        city: Madrid
        country: Spain
      properties:
        address:
          description: ""
          example: Ronda de la Comunicacion s/n
          type: string
        zipCode:
          description: ""
          example: 28050
          format: int32
          type: integer
        city:
          description: ""
          example: Madrid
          type: string
        country:
          description: ""
          example: Spain
          type: string
      required:
      - address
      - city
      - country
      - zipCode
      title: Get Attribute Value response
      type: object
    UpdateAttributeValueRequest:
      example:
        address: Ronda de la Comunicacion s/n
        zipCode: 28050
        city: Madrid
        country: Spain
      properties:
        address:
          description: ""
          example: Ronda de la Comunicacion s/n
          type: string
        zipCode:
          description: ""
          example: 28050
          format: int32
          type: integer
        city:
          description: ""
          example: Madrid
          type: string
        country:
          description: ""
          example: Spain
          type: string
      required:
      - address
      - city
      - country
      - zipCode
      title: Update Attribute Value request
      type: object
...